Injected NetworkManager using protocol, reworked NetworkManager using Alamofire. Tried to use SOLID principles
