+++ BallOfTruth/Managers/NetworkManager.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import Alamofire
 
 enum NetworkManagerError: String, Error {
     
@@ -14,51 +15,34 @@ enum NetworkManagerError: String, Error {
     case noConnection = "No internet connection, please check it"
 }
 
-class NetworkManager {
+class NetworkManager: NetworkDataProvider {
     
     static let shared = NetworkManager()
     
     private let apiURL = "https://8ball.delegator.com/magic/JSON/somequestion"
     
-    func getPrediction(completion: @escaping (Result<BallModel, NetworkManagerError>) -> Void) {
+    func fetchData(completion: @escaping (Result<BallModel, NetworkManagerError>) -> Void) {
         
         guard let url = URL(string: apiURL) else {
             completion(.failure(.somethingWentWrong))
             return
         }
         
-        let config = URLSessionConfiguration.default
-        config.requestCachePolicy = .reloadIgnoringLocalCacheData
-        config.urlCache = nil
+        AF.request(url)
+          .validate()
+          .responseDecodable(of: BallModel.self) { (response) in
+              
+              guard let _ = response.response else {
+                  completion(.failure(.somethingWentWrong))
+                  return
+              }
+              
+              guard let value = response.value else {
+                  completion(.failure(.somethingWentWrong))
+                  return
+              }
+              completion(.success(value))
+          }
 
-        let session = URLSession(configuration: config)
-        
-        session.dataTask(with: url) { data, response, error in
-            if let _ = error {
-                completion(.failure(.noConnection))
-                return
-            }
-            
-            guard let _ = response as? HTTPURLResponse else {
-                completion(.failure(.somethingWentWrong))
-                return
-            }
-            
-            
-            guard let data = data else {
-                completion(.failure(.somethingWentWrong))
-                return
-            }
-            
-            do {
-                let ball = try JSONDecoder().decode(BallModel.self, from: data)
-                DispatchQueue.main.async {
-                    completion(.success(ball))
-                }
-            } catch let error {
-                print("DEBUG: \(error.localizedDescription))")
-                completion(.failure(.unauthorized))
-            }
-        } .resume()
     }
 }
