+++ BallOfTruth/ViewControllers/MainViewController/Model/MainModel.swift
@@ -1,0 +1,12 @@
+//
+//  MainModel.swift
+//  BallOfTruth
+//
+//  Created by Илья on 22.02.2022.
+//
+
+import Foundation
+
+struct MainModel {
+    var message: String
+}
