+++ BallOfTruth/ViewControllers/MainViewController/View/MainViewController.swift
@@ -9,11 +9,13 @@ import UIKit
 
 class MainViewController: UIViewController {
     
+    // outlets
     weak private var ball: UIView!
     weak private var ballMiddle: UIView!
     weak private var messageText: UILabel!
 
-    var messageString = ""
+    // Data Provider protocol
+    var viewModel = MainViewModel()
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -25,18 +27,18 @@ class MainViewController: UIViewController {
         messageText.text = DefaultsModel.shared.welcomeString
     }
     
+    
     // when user starts to shake device app sends network request to server and text on the screen changes to one of hardcoded variants
     override func motionBegan(_ motion: UIEvent.EventSubtype, with event: UIEvent?) {
+        viewModel.getMessage()
         messageText.text = DefaultsModel.shared.shakeString
-        messageString = ""
-        loadMessage()
     }
     
     
     // after 1.5 seconds after shake shows text from json
     override func motionEnded(_ motion: UIEvent.EventSubtype, with event: UIEvent?) {
         DispatchQueue.main.asyncAfter(deadline: .now() + 1.5) {
-            self.messageText.text = self.messageString
+            self.messageText.text = self.viewModel.model?.message
         }
     }
     
@@ -44,25 +46,6 @@ class MainViewController: UIViewController {
     override func motionCancelled(_ motion: UIEvent.EventSubtype, with event: UIEvent?) {
         messageText.text = DefaultsModel.shared.shakeCancelled
     }
-    
-    private func loadMessage() {
-        NetworkManager.shared.getPrediction { [weak self] result in
-            guard self == self else { return }
-            switch result {
-            case .success(let message):
-                print(message)
-                
-                // if request was successfull result of request saves to variable
-                self?.messageString = message.magic.answer
-            case .failure(let error):
-                print(error.localizedDescription)
-                
-                // if request was unsuccessfull variable assigns hardcoded error text
-                self?.messageString = DefaultsModel.shared.shakeError
-            }
-        }
-    }
-
     
     // basic layout with constraints
     private func setupLayout() {
