+++ BallOfTruth/ViewControllers/SettingsViewController/View/Cell/SettingsCell.swift
